requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
distro==1.9.0
fake-useragent==2.2.0
Flask-Cors==6.0.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.0
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pyee==11.1.1
pyquery==2.0.1
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.4
requests-html==0.10.0
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
w3lib==2.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
